Split out InstantRange
